+++ src/query/service/src/pipelines/executor/mod.rs
@@ -18,6 +18,7 @@ mod executor_condvar;
 mod executor_graph;
 mod executor_settings;
 mod executor_tasks;
+mod executor_watch_notify;
 mod executor_worker_context;
 mod pipeline_complete_executor;
 mod pipeline_pulling_executor;
@@ -30,7 +31,7 @@ pub use executor_graph::RunningGraph;
 pub use executor_settings::ExecutorSettings;
 pub use executor_tasks::CompletedAsyncTask;
 pub use executor_tasks::ExecutorTasksQueue;
-pub use executor_tasks::WatchNotify;
+pub use executor_watch_notify::WatchNotify;
 pub use executor_worker_context::ExecutorTask;
 pub use executor_worker_context::ExecutorWorkerContext;
 pub use pipeline_complete_executor::PipelineCompleteExecutor;
